requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.35
psycopg[binary]==3.2.1
pydantic==2.8.2
requests==2.32.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator>=2.0,<3.0
